🐛 fix: 전역 모달을 close 시 unmount
