+++ app/src/@core/providers/ModalProvider.tsx
@@ -3,29 +3,15 @@ import { isSpotlightOpenAtom } from '@core/atoms/isSpotlightOpenAtom';
 import { ReLoginDialog } from '@core/components/Modal/ReLoginDialog';
 import { Spotlight } from '@core/components/Spotlight';
 import { PropsWithReactElementChildren } from '@shared/types/PropsWithChildren';
-import { useAtom } from 'jotai';
+import { useAtomValue } from 'jotai';
 
 const ModalProvider = ({ children }: PropsWithReactElementChildren) => {
-  const [isReLoginDialogOpen, setIsReLoginDialogOpen] = useAtom(
-    isReLoginDialogOpenAtom,
-  );
-  const [isSpotlightOpen, setIsSpotlightOpen] = useAtom(isSpotlightOpenAtom);
-
-  const closeReLoginDialog = () => {
-    setIsReLoginDialogOpen(false);
-  };
-
-  const closeSpotlight = () => {
-    setIsSpotlightOpen(false);
-  };
-
+  const isReLoginDialogOpen = useAtomValue(isReLoginDialogOpenAtom);
+  const isSpotlightOpen = useAtomValue(isSpotlightOpenAtom);
   return (
     <>
-      <ReLoginDialog
-        isOpen={isReLoginDialogOpen}
-        onClose={closeReLoginDialog}
-      />
-      <Spotlight isOpen={isSpotlightOpen} onClose={closeSpotlight} />
+      {isReLoginDialogOpen ? <ReLoginDialog /> : null}
+      {isSpotlightOpen ? <Spotlight /> : null}
       {children}
     </>
   );
